Move services and delete ddo folder from common.
