+++ ocean_lib/assets/credentials.py
