--- ocean_lib/common/ddo/__init__.py
@@ -1,7 +1,0 @@
-#
-# Copyright 2021 Ocean Protocol Foundation
-# SPDX-License-Identifier: Apache-2.0
-#
-
-
-"""Ocean Protocol DID/DDO Python Library"""
