--- ocean_lib/common/ddo/constants.py
@@ -1,12 +1,0 @@
-#
-# Copyright 2021 Ocean Protocol Foundation
-# SPDX-License-Identifier: Apache-2.0
-#
-"""
-Contains constant values for
-- `DID_DDO_CONTEXT_URL`
-- `PROOF_TYPE`
-"""
-DID_DDO_CONTEXT_URL = "https://w3id.org/did/v1"
-
-PROOF_TYPE = "DDOIntegritySignature"
