+++ ocean_lib/models/test/test_order.py
@@ -6,9 +6,9 @@ import os
 
 from eth_utils import remove_0x_prefix
 from ocean_lib.common.agreements.service_types import ServiceTypes
-from ocean_lib.common.ddo.service import Service
 from ocean_lib.models.data_token import DataToken
 from ocean_lib.models.order import Order
+from ocean_lib.services.service import Service
 from tests.resources.ddo_helpers import get_metadata, get_registered_ddo
