+++ ocean_lib/services/test/test_service.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 from ocean_lib.common.agreements.service_types import ServiceTypes
-from ocean_lib.common.ddo.service import Service
+from ocean_lib.services.service import Service
 from tests.resources.ddo_helpers import get_sample_ddo
